services/log-ingestion/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
pydantic==2.5.0
python-json-logger==2.0.7